requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
azure-core==1.36.0
    # via
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-storage-blob
azure-identity==1.25.1
    # via sara-thermal-reading (pyproject.toml)
azure-keyvault-secrets==4.10.0
    # via sara-thermal-reading (pyproject.toml)
azure-storage-blob==12.27.1
    # via sara-thermal-reading (pyproject.toml)
certifi==2025.11.12
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.3
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
idna==3.11
    # via requests
isodate==0.7.2
    # via
    #   azure-keyvault-secrets
    #   azure-storage-blob
loguru==0.7.3
    # via sara-thermal-reading (pyproject.toml)
msal==1.34.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
numpy==2.3.5
    # via
    #   opencv-python
    #   sara-thermal-reading (pyproject.toml)
opencv-python==4.11.0.86
    # via sara-thermal-reading (pyproject.toml)
pathlib==1.0.1
    # via sara-thermal-reading (pyproject.toml)
pillow==12.0.0
    # via sara-thermal-reading (pyproject.toml)
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via sara-thermal-reading (pyproject.toml)
pydantic-core==2.41.5
    # via pydantic
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
python-dotenv==1.2.1
    # via sara-thermal-reading (pyproject.toml)
requests==2.32.5
    # via
    #   azure-core
    #   msal
typing-extensions==4.15.0
    # via
    #   azure-core
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-storage-blob
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.1
    # via requests
